#2: Pulled process_data function to the libs folder
